+++ CHANGELOG.md
@@ -6,6 +6,28 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2017-10-27
+### Added
+- New configuration options that enable:
+  - a default interval to be set
+  - a socket heartbeat checker
+  - the ability to bring your own express app
+  - options for passing to `ws`
+  - options to set for every HTTP request
+- Source configuration property `options` for passing HTTP options to different poll types
+- Client side examples and the beginnings of a library:
+  - library source under `client/lib`
+  - includes _**.d.ts**_ type definitions under `client/lib/types` for:
+    - `xml2js`-style objects (JS from XML) + helpers for extending
+    - RSS2.0 feeds parsed with `xml2js`, made with provided XML helpers
+    - RSS extras as examples of how to further extend types with provided helpers
+  - example client source under `client/example`
+  - includes examples of how to build classes that listen to polling socket server (uses XML examples in `server/example/xml)
+
+## [1.0.1] - 2017-10-25
+### Changed
+- Switched `axios` to `rx-http-request` for a more "reactive" approach to HTTP
+
 ## [1.0.0] - 2017-10-15
 ### Added
 - First release adds `PollingSocketServer` class for setting up new servers